+++ eNMS/database.py
@@ -91,9 +91,10 @@ class Database:
             self.factory(
                 "server",
                 **{
-                    "name": str(getnode()),
-                    "description": "Localhost",
-                    "ip_address": "0.0.0.0",
+                    "name": vs.server,
+                    "description": vs.server,
+                    "mac_address": str(getnode()),
+                    "ip_address": vs.server_ip,
                     "status": "Up",
                 },
             )
@@ -307,7 +308,7 @@ class Database:
         if rbac:
             user = current_user or self.get_user(username or "admin")
             if user.is_authenticated and not user.is_admin:
-                if model in vs.rbac["admin_models"]:
+                if model in vs.rbac["admin_models"].get(rbac, []):
                     raise self.rbac_error
                 query = vs.models[model].rbac_filter(query, rbac, user)
         return query
@@ -408,7 +409,9 @@ class Database:
                     sleep(self.retry_commit_time * (index + 1))
         return instance
 
-    def get_credential(self, username, name=None, device=None, credential_type="any"):
+    def get_credential(
+        self, username, name=None, device=None, credential_type="any", optional=False
+    ):
         pool_alias = aliased(vs.models["pool"])
         query = self.session.query(vs.models["credential"])
         if device:
@@ -422,7 +425,7 @@ class Database:
         if credential_type != "any":
             query = query.filter(vs.models["credential"].role == credential_type)
         credentials = max(query.all(), key=attrgetter("priority"), default=None)
-        if not credentials:
+        if not credentials and not optional:
             raise Exception(f"No matching credentials found for DEVICE '{device.name}'")
         return credentials