+++ eNMS/models/administration.py
@@ -15,6 +15,7 @@ class Server(AbstractBase):
     id = db.Column(Integer, primary_key=True)
     name = db.Column(db.SmallString, unique=True)
     description = db.Column(db.LargeString)
+    mac_address = db.Column(db.TinyString)
     ip_address = db.Column(db.TinyString)
     weight = db.Column(Integer, default=1)
     status = db.Column(db.TinyString, default="down")
@@ -41,6 +42,9 @@ class User(AbstractBase, UserMixin):
     small_menu = db.Column(Boolean, default=False, info={"log_change": False})
     theme = db.Column(db.TinyString, default="default", info={"log_change": False})
     pools = relationship("Pool", secondary=db.pool_user_table, back_populates="users")
+    services = relationship(
+        "Service", secondary=db.service_owner_table, back_populates="owners"
+    )
     is_admin = db.Column(Boolean, default=False)
 
     def get_id(self):