+++ eNMS/services/workflow/workflow.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from heapq import heappop, heappush
-from sqlalchemy import Boolean, event, ForeignKey, Integer
-from sqlalchemy.orm import backref, relationship
+from sqlalchemy import Boolean, ForeignKey, Integer, or_
+from sqlalchemy.orm import aliased, backref, relationship
 from sqlalchemy.schema import UniqueConstraint
 
 from eNMS.database import db
@@ -45,12 +45,6 @@ class Workflow(Service):
         super().__init__(**kwargs)
         if not migration_import and self.name not in end.positions:
             end.positions[self.name] = (500, 0)
-
-    @classmethod
-    def configure_events(cls):
-        @event.listens_for(cls.services, "append")
-        def append(target, *_):
-            target.last_modified = vs.get_time()
 
     def delete(self):
         for service in self.services:
@@ -128,7 +122,7 @@ class Workflow(Service):
         tracking_bfs = run.run_method == "per_service_with_workflow_targets"
         while services:
             if run.stop:
-                return {"payload": run.payload, "success": False, "result": "Stopped"}
+                return {"payload": run.payload, "success": False, "result": "Aborted"}
             _, service = heappop(services)
             if number_of_runs[service.name] >= service.maximum_runs:
                 continue
@@ -255,6 +249,23 @@ class WorkflowEdge(AbstractBase):
     def update(self, **kwargs):
         super().update(**kwargs)
         self.set_name(kwargs.get("name"))
+
+    @classmethod
+    def rbac_filter(cls, query, mode, user):
+        originals_alias = aliased(vs.models["service"])
+        if mode == "edit":
+            query = (
+                query.join(cls.workflow)
+                .join(originals_alias, vs.models["service"].originals)
+                .join(vs.models["user"], originals_alias.owners)
+                .filter(
+                    or_(
+                        vs.models["user"].name == user.name,
+                        ~originals_alias.lock_mode.contains(mode),
+                    )
+                )
+            )
+        return query
 
     def set_name(self, name=None):
         self.name = name or f"[{self.workflow}] {vs.get_time()}"